+++ src/features/admin/errorReview/constants/errorReviewQueryKey.ts
@@ -1,0 +1,1 @@
+export const ERROR_REVIEW_QUERY_KEY = ['admin', 'errorReview', 'riskFeedbacks'] as const;
